Move `create_actions()` to a new module
Reduces clutter in `app`.
